Socket - Add new sockets subclass files.

+++ CommonSocket.cpp
@@ -49,7 +49,7 @@ void Socket::socketSend(std::string &dataToSend) {
 	bool socketIsOpen = true;
 
 	while(bytesSent < bytesToSend && socketIsOpen == true) {
-			int result = send(this->socketFD, dataToSend.c_str(), bytesSent, 0);
+			size_t result = send(this->socketFD, dataToSend.c_str(), bytesSent, 0);
 
 			if (result == kSocketError) {
 				printf("Socket send error: %sn\n", strerror(errno));
@@ -74,7 +74,7 @@ std::string Socket::socketRecieve(size_t dataToRecieveSize) {
 	    bool socketOpen = true;
 
 	    while (dataReceivedSize < dataToRecieveSize && socketOpen == true) {
-	        int result;
+	        size_t result;
 	        std::vector<char> buffer;
 	        buffer.resize(dataToRecieveSize - dataReceivedSize, 0);
